fix: use `size_t` for output of `sizeof()` + small tweaks

+++ include/sk_rngbuf.h
@@ -21,13 +21,14 @@
 
 #pragma once
 
+#include <stddef.h>
 #include <sk_defines.h>
 
 #define SK_RNGBUF_AT(rb, i) ((void *) ((i32 *) (rb)->data + ((((rb)->head + (i)) % (rb)->capacity) * (rb)->element_size)))
 
 typedef struct {
   u32 capacity;
-  u32 element_size;
+  size_t element_size;
   u8 overwrite;
   u32 curr_len;
   i32 head;
